+++ CHANGES.md
@@ -100,6 +100,12 @@
 
 ### Bug Fixes and Minor Changes
 
+  * `XMonad.Hooks.StatusBar.PP`
+
+    - `dynamicLogString` now forces its result and produces an error string if
+      it throws an exception. Use `dynamicLogString'` if for some reason you
+      need the old behavior.
+
   * `XMonad.Util.EZConfig`
 
     - Added `remapKeysP`, which remaps keybindings from one binding to